pyproject.toml
[project]
name = "bitpeer"
version = "0.1.0"
description = "Local-first Bybit P2P analytics for RUB/USDT and VND/USDT."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "duckdb>=1.0.0",
  "httpx>=0.27.0",
  "polars>=0.20.0",
  "pydantic>=2.7.0",
  "pyarrow>=16.0.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.0",
  "tenacity>=8.3.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dashboard = [
  "pandas>=2.2.0",
  "plotly>=5.22.0",
  "streamlit>=1.35.0",
]
dev = [
  "pyright>=1.1.0",
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
bitpeer = "bitpeer.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bitpeer"]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
